Increase Z_PROBE_OFFSET_FROM_EXTRUDER
Using -.18 ends up with a head crash. It's better to let users adjust their
G29 than risk a head crash if they don't use a negative adjustment for G29.

For what it's worth, the original value of this was -17.65 but my printer
works perfectly with `G29 Z0.2` so I rounded up for a simpler number.