Enable thermal runaway protection
From Tyler A at MatterHackers:

We did some experimentation to find the best settings by pulling the thermistor
out once it was at temperature and measuring how hot it got before it shut off
the heater. In 10 seconds it only got about 30 degrees above the set value.

We also tested to make sure these values wouldn't trigger any false positives.
A hysteresis of 4 degrees gives it plenty of leeway, even with bad PID tuning.
When we cooled the hotend with a fan after it was at temp, it didn't go out of
bounds long enough to trigger the protection.

We haven't tested the settings for the bed. Those are a best guess.

Note that this protection only kicks in once the hotend has reached operating
temperature. You are still at risk while its warming up.